[main] REF-8 Move the connect method into town-controller
